Rename lib to dist because I keep accidentally opening lib.
