+++ examples/bot-conversation.js
@@ -11,7 +11,7 @@ const chalk = require('chalk');
 //   import {createBot, createConversation, createMatcher, createParser} from 'chatter';
 // ES5 syntax:
 //   const chatter = require('chatter');
-const chatter = require('../lib');
+const chatter = require('..');
 const createBot = chatter.createBot;
 const createConversation = chatter.createConversation;
 const createMatcher = chatter.createMatcher;
