+++ examples/slack-naive.js
@@ -12,7 +12,7 @@
 //   import {processMessage, createMatcher} from 'chatter';
 // ES5 syntax:
 //   const chatter = require('chatter');
-const chatter = require('../lib');
+const chatter = require('..');
 const processMessage = chatter.processMessage;
 const createMatcher = chatter.createMatcher;
